fix: rename file with typo
